+++ components/ui/BannerGrid.tsx
@@ -102,8 +102,8 @@ export default function BannnerGrid({
         )}
       <div
         class={`grid gap-4 md:gap-6 ${
-          MOBILE_COLUMNS[itemsPerLine.mobile ?? 2]
-        } ${DESKTOP_COLUMNS[itemsPerLine.desktop ?? 4]}`}
+          MOBILE_COLUMNS[itemsPerLine?.mobile ?? 2]
+        } ${DESKTOP_COLUMNS[itemsPerLine?.desktop ?? 4]}`}
       >
         {banners.map(({ href, srcMobile, srcDesktop, alt }) => (
           <a
