+++ components/ui/Features.tsx
@@ -26,7 +26,7 @@ function FeatureHighlights(
     <div class="container min-h-[280px] p-6 sm:px-0 sm:py-10">
       <div class="border-base-200 border border-solid">
         <div class="flex flex-col justify-evenly divide-y divide-base-200 mx-6 sm:flex-row sm:divide-y-0 sm:divide-x sm:mx-0 sm:my-10">
-          {features.map(({ icon: id = "Truck", title, description }) => (
+          {features?.map(({ icon: id = "Truck", title, description }) => (
             <div class="flex flex-row gap-4 py-6 sm:flex-col sm:py-0 sm:px-8">
               <Icon
                 id={id}
