+++ import_map.json
@@ -2,8 +2,8 @@
   "imports": {
     "deco-sites/fashion/": "./",
     "$store/": "./",
-    "deco-sites/std/": "https://denopkg.com/deco-sites/std@1.9.19/",
-    "$live/": "https://denopkg.com/deco-cx/live@1.9.0/",
+    "deco-sites/std/": "https://denopkg.com/deco-sites/std@1.9.23/",
+    "$live/": "https://denopkg.com/deco-cx/live@1.9.3/",
     "$fresh/": "https://denopkg.com/deco-cx/fresh@1.2.0/",
     "preact": "https://esm.sh/preact@10.13.2",
     "preact/": "https://esm.sh/preact@10.13.2/",
