+++ live.gen.ts
@@ -18,20 +18,33 @@ import * as $$$$$8 from "./islands/ProductImageZoom.tsx";
 import * as $$$$$9 from "./islands/OutOfStock.tsx";
 import * as $$$$$10 from "./islands/SearchControls.tsx";
 import * as $$$$$$$$0 from "./sections/WishlistGallery.tsx";
-import * as $$$$$$$$1 from "./sections/LinkTree.tsx";
-import * as $$$$$$$$2 from "./sections/Features.tsx";
-import * as $$$$$$$$3 from "./sections/BannerPLP.tsx";
-import * as $$$$$$$$4 from "./sections/BannerGrid.tsx";
-import * as $$$$$$$$5 from "./sections/SearchResult.tsx";
-import * as $$$$$$$$6 from "./sections/ProductShelf.tsx";
-import * as $$$$$$$$7 from "./sections/Footer.tsx";
-import * as $$$$$$$$8 from "./sections/CookieConsent.tsx";
-import * as $$$$$$$$9 from "./sections/DesignSystem.tsx";
-import * as $$$$$$$$10 from "./sections/Header.tsx";
-import * as $$$$$$$$11 from "./sections/ProductDetails.tsx";
-import * as $$$$$$$$12 from "./sections/Highlights.tsx";
-import * as $$$$$$$$13 from "./sections/WhatsApp.tsx";
-import * as $$$$$$$$14 from "./sections/Carousel.tsx";
+import * as $$$$$$$$1 from "./sections/DesignSystem/DesignSystem.tsx";
+import * as $$$$$$$$2 from "./sections/LinkTree.tsx";
+import * as $$$$$$$$3 from "./sections/Features.tsx";
+import * as $$$$$$$$4 from "./sections/BannerPLP.tsx";
+import * as $$$$$$$$5 from "./sections/Footers/Footer.tsx";
+import * as $$$$$$$$6 from "./sections/BannerGrid.tsx";
+import * as $$$$$$$$7 from "./sections/SearchResult.tsx";
+import * as $$$$$$$$8 from "./sections/Product/WishlistGallery.tsx";
+import * as $$$$$$$$9 from "./sections/Product/Details.tsx";
+import * as $$$$$$$$10 from "./sections/Product/SearchResult.tsx";
+import * as $$$$$$$$11 from "./sections/Product/Shelf.tsx";
+import * as $$$$$$$$12 from "./sections/Miscellaneous/LinkTree.tsx";
+import * as $$$$$$$$13 from "./sections/Miscellaneous/CookieConsent.tsx";
+import * as $$$$$$$$14 from "./sections/Miscellaneous/WhatsApp.tsx";
+import * as $$$$$$$$15 from "./sections/ProductShelf.tsx";
+import * as $$$$$$$$16 from "./sections/Banners/Grid.tsx";
+import * as $$$$$$$$17 from "./sections/Banners/PLP.tsx";
+import * as $$$$$$$$18 from "./sections/Banners/Carousel.tsx";
+import * as $$$$$$$$19 from "./sections/Footer.tsx";
+import * as $$$$$$$$20 from "./sections/CookieConsent.tsx";
+import * as $$$$$$$$21 from "./sections/DesignSystem.tsx";
+import * as $$$$$$$$22 from "./sections/Headers/Header.tsx";
+import * as $$$$$$$$23 from "./sections/Header.tsx";
+import * as $$$$$$$$24 from "./sections/ProductDetails.tsx";
+import * as $$$$$$$$25 from "./sections/Highlights.tsx";
+import * as $$$$$$$$26 from "./sections/WhatsApp.tsx";
+import * as $$$$$$$$27 from "./sections/Carousel.tsx";
 import * as $live_middleware from "$live/routes/_middleware.ts";
 import * as $live_workbench from "$live/routes/live/workbench.ts";
 import * as $live_invoke from "$live/routes/live/invoke/index.ts";
@@ -176,20 +189,33 @@ const manifest = {
     "$live/sections/PageInclude.tsx": i2$$$$0,
     "$live/sections/Slot.tsx": i2$$$$1,
     "$live/sections/UseSlot.tsx": i2$$$$2,
-    "deco-sites/fashion/sections/BannerGrid.tsx": $$$$$$$$4,
-    "deco-sites/fashion/sections/BannerPLP.tsx": $$$$$$$$3,
-    "deco-sites/fashion/sections/Carousel.tsx": $$$$$$$$14,
-    "deco-sites/fashion/sections/CookieConsent.tsx": $$$$$$$$8,
-    "deco-sites/fashion/sections/DesignSystem.tsx": $$$$$$$$9,
-    "deco-sites/fashion/sections/Features.tsx": $$$$$$$$2,
-    "deco-sites/fashion/sections/Footer.tsx": $$$$$$$$7,
-    "deco-sites/fashion/sections/Header.tsx": $$$$$$$$10,
-    "deco-sites/fashion/sections/Highlights.tsx": $$$$$$$$12,
-    "deco-sites/fashion/sections/LinkTree.tsx": $$$$$$$$1,
-    "deco-sites/fashion/sections/ProductDetails.tsx": $$$$$$$$11,
-    "deco-sites/fashion/sections/ProductShelf.tsx": $$$$$$$$6,
-    "deco-sites/fashion/sections/SearchResult.tsx": $$$$$$$$5,
-    "deco-sites/fashion/sections/WhatsApp.tsx": $$$$$$$$13,
+    "deco-sites/fashion/sections/BannerGrid.tsx": $$$$$$$$6,
+    "deco-sites/fashion/sections/BannerPLP.tsx": $$$$$$$$4,
+    "deco-sites/fashion/sections/Banners/Carousel.tsx": $$$$$$$$18,
+    "deco-sites/fashion/sections/Banners/Grid.tsx": $$$$$$$$16,
+    "deco-sites/fashion/sections/Banners/PLP.tsx": $$$$$$$$17,
+    "deco-sites/fashion/sections/Carousel.tsx": $$$$$$$$27,
+    "deco-sites/fashion/sections/CookieConsent.tsx": $$$$$$$$20,
+    "deco-sites/fashion/sections/DesignSystem.tsx": $$$$$$$$21,
+    "deco-sites/fashion/sections/DesignSystem/DesignSystem.tsx": $$$$$$$$1,
+    "deco-sites/fashion/sections/Features.tsx": $$$$$$$$3,
+    "deco-sites/fashion/sections/Footer.tsx": $$$$$$$$19,
+    "deco-sites/fashion/sections/Footers/Footer.tsx": $$$$$$$$5,
+    "deco-sites/fashion/sections/Header.tsx": $$$$$$$$23,
+    "deco-sites/fashion/sections/Headers/Header.tsx": $$$$$$$$22,
+    "deco-sites/fashion/sections/Highlights.tsx": $$$$$$$$25,
+    "deco-sites/fashion/sections/LinkTree.tsx": $$$$$$$$2,
+    "deco-sites/fashion/sections/Miscellaneous/CookieConsent.tsx": $$$$$$$$13,
+    "deco-sites/fashion/sections/Miscellaneous/LinkTree.tsx": $$$$$$$$12,
+    "deco-sites/fashion/sections/Miscellaneous/WhatsApp.tsx": $$$$$$$$14,
+    "deco-sites/fashion/sections/Product/Details.tsx": $$$$$$$$9,
+    "deco-sites/fashion/sections/Product/SearchResult.tsx": $$$$$$$$10,
+    "deco-sites/fashion/sections/Product/Shelf.tsx": $$$$$$$$11,
+    "deco-sites/fashion/sections/Product/WishlistGallery.tsx": $$$$$$$$8,
+    "deco-sites/fashion/sections/ProductDetails.tsx": $$$$$$$$24,
+    "deco-sites/fashion/sections/ProductShelf.tsx": $$$$$$$$15,
+    "deco-sites/fashion/sections/SearchResult.tsx": $$$$$$$$7,
+    "deco-sites/fashion/sections/WhatsApp.tsx": $$$$$$$$26,
     "deco-sites/fashion/sections/WishlistGallery.tsx": $$$$$$$$0,
     "deco-sites/std/sections/Analytics.tsx": i2$$$$3,
     "deco-sites/std/sections/configButterCMS.global.tsx": i2$$$$4,
