+++ sections/Banners/PLP.tsx
@@ -1,0 +1,1 @@
+export { default } from "$store/components/ui/BannerPLP.tsx";
