+++ sections/Footers/Footer.tsx
@@ -1,0 +1,1 @@
+export { default } from "$store/components/footer/Footer.tsx";
