+++ sections/Headers/Header.tsx
@@ -1,0 +1,1 @@
+export { default } from "$store/components/header/Header.tsx";
