+++ sections/Miscellaneous/CookieConsent.tsx
@@ -1,0 +1,1 @@
+export { default } from "$store/components/ui/CookieConsent.tsx";
