+++ sections/Product/SearchResult.tsx
@@ -1,0 +1,1 @@
+export { default } from "$store/components/search/SearchResult.tsx";
