+++ sections/Product/WishlistGallery.tsx
@@ -1,0 +1,1 @@
+export { default } from "$store/components/wishlist/WishlistGallery.tsx";
